Move code action stuff into code action provider
