+++ src/providers/diagnostics/typeInferenceDiagnostics.ts
@@ -91,90 +91,4 @@ export class TypeInferenceDiagnostics {
       );
     });
   }
-
-  public onCodeAction(params: CodeActionParams): CodeAction[] {
-    const { uri } = params.textDocument;
-    const typeInferenceDiagnostics: Diagnostic[] = this.filterTypeInferenceDiagnostics(
-      params.context.diagnostics,
-    );
-
-    return this.convertDiagnosticsToCodeActions(typeInferenceDiagnostics, uri);
-  }
-
-  private filterTypeInferenceDiagnostics(
-    diagnostics: Diagnostic[],
-  ): Diagnostic[] {
-    return diagnostics.filter(
-      (diagnostic) => diagnostic.source === this.TYPE_INFERENCE,
-    );
-  }
-
-  private convertDiagnosticsToCodeActions(
-    diagnostics: Diagnostic[],
-    uri: string,
-  ): CodeAction[] {
-    const result: CodeAction[] = [];
-
-    const elmWorkspace = this.elmWorkspaceMatcher.getElmWorkspaceFor(
-      URI.parse(uri),
-    );
-
-    const forest = elmWorkspace.getForest();
-    const treeContainer = forest.getByUri(uri);
-    const checker = elmWorkspace.getTypeChecker();
-
-    if (treeContainer) {
-      diagnostics.forEach((diagnostic) => {
-        switch (diagnostic.code) {
-          case "missing_type_annotation":
-            {
-              const nodeAtPosition = TreeUtils.getNamedDescendantForPosition(
-                treeContainer.tree.rootNode,
-                diagnostic.range.start,
-              );
-
-              if (nodeAtPosition.parent) {
-                const typeString: string = checker.typeToString(
-                  checker.findType(nodeAtPosition.parent),
-                  treeContainer,
-                );
-
-                result.push(
-                  this.insertQuickFixAtStart(
-                    uri,
-                    `${nodeAtPosition.text} : ${typeString}\n`,
-                    diagnostic,
-                    "Add inferred annotation",
-                  ),
-                );
-              }
-            }
-            break;
-        }
-      });
-    }
-
-    return result;
-  }
-
-  private insertQuickFixAtStart(
-    uri: string,
-    replaceWith: string,
-    diagnostic: Diagnostic,
-    title: string,
-  ): CodeAction {
-    const map: {
-      [uri: string]: TextEdit[];
-    } = {};
-    if (!map[uri]) {
-      map[uri] = [];
-    }
-    map[uri].push(TextEdit.insert(diagnostic.range.start, replaceWith));
-    return {
-      diagnostics: [diagnostic],
-      edit: { changes: map },
-      kind: CodeActionKind.QuickFix,
-      title,
-    };
-  }
 }
